Remove test code about System.IdentityModel.CanonicalizationDriver.
